+++ airbyte-cdk/java/airbyte-cdk/db-sources/src/testFixtures/kotlin/io/airbyte/cdk/integrations/standardtest/source/AbstractSourceDatabaseTypeTest.kt
@@ -132,7 +132,7 @@ abstract class AbstractSourceDatabaseTypeTest : AbstractSourceConnectorTest() {
      */
     @Test
     @Throws(Exception::class)
-    fun testDataContent() {
+    open fun testDataContent() {
         // Class used to make easier the error reporting
         class MissedRecords( // Stream that is missing any value
             var streamName:
@@ -148,7 +148,7 @@ abstract class AbstractSourceDatabaseTypeTest : AbstractSourceConnectorTest() {
         val recordMessages =
             allMessages!!
                 .stream()
-                .filter { m: AirbyteMessage? -> m!!.type == AirbyteMessage.Type.RECORD }
+                .filter { m: AirbyteMessage -> m.type == AirbyteMessage.Type.RECORD }
                 .toList()
         val expectedValues: MutableMap<String?, MutableList<String?>?> = HashMap()
         val missedValuesByStream: MutableMap<String?, ArrayList<MissedRecords>> = HashMap()
@@ -256,7 +256,7 @@ abstract class AbstractSourceDatabaseTypeTest : AbstractSourceConnectorTest() {
      * scripts failed.
      */
     @Throws(Exception::class)
-    protected fun createTables() {
+    protected open fun createTables() {
         for (test in testDataHolders) {
             database!!.query<Any?> { ctx: DSLContext? ->
                 ctx!!.fetch(test.createSqlQuery)
@@ -267,7 +267,7 @@ abstract class AbstractSourceDatabaseTypeTest : AbstractSourceConnectorTest() {
     }
 
     @Throws(Exception::class)
-    protected fun populateTables() {
+    protected open fun populateTables() {
         for (test in testDataHolders) {
             database!!.query<Any?> { ctx: DSLContext? ->
                 test.insertSqlQueries.forEach(Consumer { sql: String? -> ctx!!.fetch(sql) })