+++ airbyte-integrations/connectors/source-microsoft-teams/source_microsoft_teams/__init__.py
@@ -1,3 +1,8 @@
+#
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+#
+
+
 from .source import SourceMicrosoftTeams
 
 __all__ = ["SourceMicrosoftTeams"]